requirements.txt
pandas
groq==0.5.0
streamlit==1.31.0
python-dotenv==1.0.0
matplotlib==3.8.2
plotly==5.18.0
tabulate==0.9.0
